Reorganized code in lieu of removing ip-reconciler binary, and (in theory) ip_test.go should remain functional
