+++ pkg/controlloop/entity_generators.go
