+++ pkg/reconciler/errors.go
@@ -1,4 +1,3 @@
-// package main
 package reconciler
 
 const (
